+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+v 0.2.0-beta
+Fixes #11. Relationships must now be defined on the model and Resource to be
+assigned. To-many relationships can now be set to `null` without error.
+
 v 0.1.1-beta
 Fixes an issue with comparing strings to symbols when building the list of
 assignable attributes.